Updates Serverspec tests for localhost/hostname check
The Serverspec tests were inappropriately checking for the machine
hostname (i.e. localhost) to be defined as "127.0.1.1", rather than
"127.0.0.1". According to the manpage for `hosts`, 127.0.1.1 is typically
used for an FQDN, which isn't applicable in the context of SecureDrop,
since the servers are only accessible via Tor Hidden Services.

The Ubuntu Vagrant boxes built by Canonical showed the tests passing,
but since switching to Bento boxes, the test started failing with
vagrant.vm being set as the FQDN for the VM.