Creates securedrop-keyring deb package directory structure
Simple as it gets: store a binary keyring export in the directory
structure of apt's keyrings. The existing build package roles in the
securedrop repo handle bottling up the files in a deb package just fine.

The keyring file was generated using:

  gpg --no-default-keyring --keyring
  install_files/securedrop-keyring/etc/apt/trusted.gpg.d/securedrop-keyring.gpg
  --import
  install_files/ansible-base/roles/install_fpf_repo/files/fpf-signing-key.pub

WARNING: Currently using a TESTING key during development.